+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Changes
 
 * Fix Null reference on SentryXamarinFormsIntegration.Register. (#58) @lucas-zimerman
+* Add android unhandled exception and native crash sample. (#56) @lucas-zimerman
 
 ## 1.0.2